fix errors of Linux-build: client emptyness calls

+++ google/google_port.h
@@ -127,12 +127,12 @@ typedef unsigned __int64 uint64;
 typedef int8_t int8;
 typedef int16_t int16;
 typedef int32_t int32;
-typedef int64_t int64;
+typedef long long int int64;
 
 typedef uint8_t uint8;
 typedef uint16_t uint16;
 typedef uint32_t uint32;
-typedef uint64_t uint64;
+typedef unsigned long long int uint64;
 #endif
 
 // long long macros to be used because gcc and vc++ use different suffixes,
